changed direction/coordinate system to anti-clockwise==positive, east==0

+++ Board/BoardMetaTemplates.h
@@ -19,10 +19,10 @@ class INIT_MAN_JUMP_GROUP;
 template<typename T, size_t, size_t = T::SQUARE_RANGE - 1>
 class INIT_JUMPABLE;
 
-template<typename, size_t>
+template<typename, int>
 class SQUARE2BIT;
 
-template<typename, size_t>
+template<typename, int>
 class BIT2SQUARE;
 
 // include template definitions inside header since "export" keyword is not supported by Visual C++
