+++ Board/GridPredicates.h
@@ -1,13 +1,13 @@
 #pragma once
 #include "../Utilities/IntegerTypes.h"
 
-template<typename, size_t>
+template<typename, int>
 class SQUARE_IS_VALID;
 
-template<typename, size_t>
+template<typename, int>
 class SQUARE2COORD;
 
-template<typename, size_t, size_t>
+template<typename, int, int>
 class COORD2SQUARE;
 
 // include template definitions inside header since "export" keyword is not supported by Visual C++
